Bug# 1614 Switching to double as the default floating point storage to match Java.  Only storing values as long if they must be stored that way.

+++ Etomo/src/etomo/comscript/ConstTiltParam.java
@@ -137,13 +137,13 @@ public interface ConstTiltParam extends CommandDetails {
 
   public String getLogShift();
 
-  public float getRadialBandwidth();
+  public double getRadialBandwidth();
 
-  public float getRadialFalloff();
+  public double getRadialFalloff();
 
-  public float getScaleCoeff();
+  public double getScaleCoeff();
 
-  public float getScaleFLevel();
+  public double getScaleFLevel();
 
   public int getThickness();
 
@@ -153,7 +153,7 @@ public interface ConstTiltParam extends CommandDetails {
 
   public double getXAxisTilt();
 
-  public float getXShift();
+  public double getXShift();
 
   public ConstEtomoNumber getZShift();
